More new stuff, but...
...we generate way too much data on disk. MAUI hello world has 111
assemblies and, after linking and compression, they result in over 130mb
of .ll sources.  The multiplication factor for dll to ll conversion is 9
characters per byte, because of the format used in LLVM IR arrays.
That's way too much data written to disk, `obj` is big enough as it is.

The solution to implement tomorrow - don't flush MemoryStream we use to
keep the generated source to disk, but rather pipe it to `llc` via
S.D.Process support for stdin redirection.  This is fine for this task
because we're not using the .ll sources in msbuild targets as inputs or
outputs, we're only interested in the .so files the task produces.

MemoryStream contents will be dumped to disk only if an MSBuild
variable (name to be chosen) is set to `true` or when `llc` fails to
compile the source.