[Xamarin.Android.Build.Tasks] Make manifestmerger.jar the default.
Context https://github.com/xamarin/xamarin-android/issues/8387

The value for `AndroidManifestMerger` should be updated to
be `manifestmerger.jar`. This allows users to make use of
the new features like `AndroidManifestOverlay` out of the box.

The old system is still available, the users will need to
add the following to their csproj.

```
<AndroidManifestMerger>legacy</AndroidManifestMerger>
```