[Xamarin.Android.Build.Tasks] case-insensitive Legacy Designer fixup (#8376)
Context: 628d6cb869c872c34d17b014724a43e8d451fb93
Context: dc3ccf28cdbe9f8c0a705400b83c11a85c81a980

We have a very odd situation with the
[`Xamarin.CommunityToolkit.MauiCompat` NuGet package][0]:

The NuGet package ships with a file `.net/__res_name_case_map.txt`
(628d6cb8) in `Xamarin.CommunityToolkit.MauiCompat.aar` which contains
the case map changes for the `@(AndroidResource)` items.  The purpose
of this file is to allow the user to use PascalCase (or any case) names
in C# code, while the Android Resource compilation is left as all
lowercase.  This is because Android requires that file-backed resource
names consist of only lowercase letters.

The contents of this file includes:

	Resources/Layout/CameraFragment.axml;layout/camerafragment.xml

which tells us that the Android `R.layout.camerafragment` resource id
should be "bound" as `Resouce.Layout.CameraFragment`, which allows
[`CameraFragment.android.cs` to build][1]:

	public override AView? OnCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) =>
	    inflater.Inflate(Resource.Layout.CameraFragment, null);

However, when we look at the IL that comes from the NuGet package we
see the only lowercase `camerafragment`:

	ikdasm ~/.nuget/packages/xamarin.communitytoolkit.mauicompat/2.0.2-preview1013/lib/net6.0-android31.0/Xamarin.CommunityToolkit.MauiCompat.dll| grep Layout::camerafragment
	…
	    IL_0131:  stsfld     int32 Xamarin.CommunityToolkit.MauiCompat.Resource/Layout::camerafragment
	…
	    IL_0001:  ldsfld     int32 Xamarin.CommunityToolkit.MauiCompat.Resource/Layout::camerafragment

For some reason as part of the XamarinCommunityToolkit build
[the following script][2] was run on the code:

	sed -i '' 's/MauiCompat.Resource.Layout.CameraFragment/MauiCompat.Resource.Layout.camerafragment/g' ./src/CommunityToolkit/Xamarin.CommunityToolkit.MauiCompat/**/CameraFragment.android.cs

This changes `Resource.Layout.CameraFragment` to
`Resource.Layout.camerafragment`, apparently because *something* is
renaming [`CameraFragment.axml`][3] to `camerafragment.axml` *after*
`__res_name_case_map.txt` is generated, but before
`CameraFragment.android.cs` is compiled.
The NuGet still includes the `.aar` with the case map changes, which
are now inconsistent with the assembly contents.

I'm not sure why XamarinCommunityToolkit does all this.

The result of the above XamarinCommunityToolkit build and packaging
behavior is that it interacts badly with the new net8.0-android
Resource Designer Assembly (dc3ccf28) work, with a build-time error:

	error XA8000: Could not find Android Resource ‘@layout/camerafragment’. Please update @(AndroidResource) to add the missing resource.

This is because we are looking for the lowercase version of the
resource in the Designer assembly property list, but because of the
remap file we only know about the PascalCase version. 

As a workaround, fallback to a *case-insensitive* resource name
lookup if a case-sensitive lookup fails.

[0]: https://www.nuget.org/packages/Xamarin.CommunityToolkit.MauiCompat/
[1]: https://github.com/xamarin/XamarinCommunityToolkit/blob/5a6062f3c3543acda3c36ca4683cd8fc7fe86ba7/src/CommunityToolkit/Xamarin.CommunityToolkit/Views/CameraView/Android/CameraFragment.android.cs#L129
[2]: https://github.com/xamarin/XamarinCommunityToolkit/blob/5a6062f3c3543acda3c36ca4683cd8fc7fe86ba7/MauiCompat.sh#L561
[3]: https://github.com/xamarin/XamarinCommunityToolkit/blob/5a6062f3c3543acda3c36ca4683cd8fc7fe86ba7/src/CommunityToolkit/Xamarin.CommunityToolkit/Resources/Layout/CameraFragment.axml