Hush some messages logged by default on startup
We log a number of messages during startup, whether or not logging is
enabled.  However, logcat messages take non-trivial (and not entirely
predictable) amount of time to process, so we should limit their number
by default in order to speed up startup.

Tests show the following improvements resulting from this change:

  * Displayed: 1.49% faster
  * Native-to-managed transition: 7.47% faster
  * Total native init time: 5.94% faster