[WIP] attempt to fix dotnet/maui#17265
Context: https://github.com/dotnet/maui/issues/17265

Reproduced the issue in a new `FixLegacyResourceDesignerStep` test. It
crashes at runtime, but if you change the class libraries in the test to
`net8.0`, it works fine.

I first attempted to fix the problem for `Release` mode, as it seemed
easier to do a "second pass" on *all assemblies* at the end of a linker
step.

This change starts to work:

    ILLink:    Library2 has an assembly reference to Library1, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null
    ILLink:    Library2 has an assembly reference with a designer.
    ILLink:     Adding reference _Microsoft.Android.Resource.Designer.
    ILLink:     FixupAssemblyTypes Library2.
    ILLink: Looking for String::hello.
    ILLink:     Fixing up System.Int32 Foo::get_Hello()
    ILLink:       Replacing IL_0000: ldsfld System.Int32 Library1.Resource/String::hello
    ILLink:       With IL_0000: call System.Int32 _Microsoft.Android.Resource.Designer.Resource/String::get_hello()

But then fails with:

    System.ArgumentException: Member 'System.Int32 _Microsoft.Android.Resource.Designer.Resource/String::get_hello()' is declared in another module and needs to be imported