Some experiments
Will most likely end up using function pointers instead of virtual
dispatch.  A bit more cumbersome but more efficient.

The goal is to get rid of `if` statements which check, for each assembly
load request, whether we should load from the APK or from the filesystem
- this is simply wasted time, since we know at the start which one of
those choices is correct and that it doesn't change until the app is
reinstalled.