[Xamarin.Android.Build.Tasks] Include Bindings Design Time builds (#8357)
Fixes: https://github.com/xamarin/xamarin-android/issues/7692

When `.java` files are part of a project through
`@(AndroidJavaSource)`, it would be nice if the *bindings* for those
Java types would be available through IDE IntelliSense/code completion.

Update the `AddBindingsToCompile` target so that it includes all
`$(IntermediateOutputPath)generated\**\*.cs` files into the design-
time build.

Additionally, fix an issue where if you remove the libraries the
generated files are not removed.

Finally, move the Layout CodeBehind generated files to their own
`$(IntermediateOutputPath)codebehind` directory.  This is to keep the
code behind and generated code separate, as it makes it easier to
update and clean up if one or the other changes.  When all the files
are in the same location we might end up deleting files we otherwise
need during an update process, slowing down the build.