[profiled-aot] update profile for .NET 8 GA
First, I had to update the URL we test, because it was down:

https://httpstat.us/200

Instead, I used:

https://httpbin.org/status/200

Next, after reviewing the latest methods in the profile, I noticed a lot
of timezone logic was completely removed. This might be related to
different cookies returns from the new URL, not sure.

I added a `DateTime.Now` call to `CommonMethods.cs`, to ensure the
timezone logic is in the AOT profile.

Testing a `dotnet new android` app on a Pixel 5, an average of 10 runs
of the "displayed time`:

    Average(ms): 179.1
    Std Err(ms): 2.53179778023443
    Std Dev(ms): 8.00624756049924