[docs] Don't overwrite xml produced by docs build (#8451)
The `mdoc export-msxdoc` command should only be used to update the
shipping `Mono.Android.xml` file during the regular build.  We do want
to run it during the API docs build for validation purposes as well, but
it should not overwrite the output of the API docs build.  We should
preserve the additions to `Mono.Android.xml` that were made during the
API docs build to update xamarin/android-api-docs.