[Xamarin.Android.Build.Tasks] Add support for @(AndroidMavenLibrary). (#8420)
Context: https://github.com/xamarin/xamarin-android/issues/4528

Add support for the `@(AndroidMavenLibrary)` item group, which will
download the requested Java artifact version from a Maven repository
and add it as an `@(AndroidLibrary)` for binding.

For example, to download the Maven package
[pkg:maven/com.google.auto.value/auto-value-annotations@1.10.4][0]
from Maven Central and create an Android binding for it:

	<ItemGroup>
	  <AndroidMavenLibrary Include="com.google.auto.value:auto-value-annotations" Version="1.10.4" />
	</ItemGroup>

~~ Specification ~~

The `//AndroidMavenLibrary/@Include` format is `{GroupId}:{ArtifactId}`.

Any additional attributes such as `%(Bind)` or `%(Pack)` will be copied
to the created `@(AndroidLibrary)` item.

`%(Bind)` and `%(Pack)` default to `true` and control the binding
process as specified for [`@(AndroidLibrary)`][1].

`%(Repository)` controls which Maven Repository to download artifacts
from.  Supported values include:

  * `Central` for [Maven Central][2].  This is the default value if
    `%(Repository)` is not specified.

  * `Google` for [Google's Maven][3].

  * A URL, for downloading from a custom Maven instance.
    *Note*: Maven authentication is *not* supported.

~~Examples:~~

	<ItemGroup>
	  <AndroidMavenLibrary 
	      Include="androidx.core:core" 
	      Version="1.9.0" 
	      Repository="Google"
	      Bind="false"
	  />
	  <AndroidMavenLibrary 
	      Include="com.github.chrisbanes:PhotoView" 
	      Version="2.3.0" 
	      Repository="https://repository.mulesoft.org/nexus/content/repositories/public"
	      Pack="false"
	  />
	</ItemGroup>

There are 2 parts to #4528:

 1. Downloading artifacts from Maven
 2. Ensuring all dependencies specified in the POM file are met

Only (1) is currently addressed. (2) will be addressed in the future.

[0]: https://repo1.maven.org/maven2/com/google/auto/value/auto-value-annotations/1.10.4/
[1]: https://github.com/xamarin/xamarin-android/blob/main/Documentation/guides/OneDotNetEmbeddedResources.md#msbuild-item-groups
[2]: https://repo1.maven.org/maven2/
[3]: https://maven.google.com/web/index.html