[Microsoft.Android.Templates] docs link in binding project template
Fixes: https://github.com/xamarin/xamarin-android/issues/8411
Context: https://github.com/dotnet/docs-maui/pull/1916

We have a nice docs page that should eliminate some confusion if we
linked to it with a comment in the project template.

I considered putting a commented out example of disabling the C# binding
(ala `EmbeddedReferenceJar`):

    <ItemGroup>
      <AndroidLibrary Update="foo.jar" Bind="false">
    </ItemGroup

But I think this case is rare, and the docs page should be sufficient.