[tests] trim trailing `\` from AndroidSdkDirectory/JavaSdkDirectory (#8564)
When running MSBuild tests locally, both of the paths used for:

    /p:AndroidSdkDirectory="C:\Some Path\android-sdk\"
    /p:JavaSdkDirectory="C:\Some Path\jdk\"

Because the two paths have a trailing `\`, it appears that MSBuild is
escaping the `\"` giving an error message that we can't build more than
one project! Basically, all command-line parsing is failing!

I am unsure what changed, but making this change locally allows me to
run the MSBuild tests again.