[Xamarin.Android.Build.Tasks] introduce `XA1038`
Fixes: https://github.com/xamarin/xamarin-android/issues/8331

Building a `net8.0-android33` project, currently fails with:

    error NETSDK1181: Error getting pack version: Pack 'Microsoft.Android.Ref.33' was not present in workload manifests.
    C:\Program Files\dotnet\sdk\8.0.100-preview.7.23376.3\Sdks\Microsoft.NET.Sdk\targets\Microsoft.NET.Sdk.FrameworkReferenceResolution.targets

To solve this, you would either change 33 to 34, or just remove the
number to rely on the default value.

To make this easier:

* Automatically switch to 34 if the user specifies 33 or less.

* Emit the new `XA1038` error message.

This allows these projects to build with a reasonable error message.

The only concern down the road: if we ever support `net8.0-android35`
alongside `net8.0-android34`, then we'd need to slightly adjust the
logic here.