Some test fixes
Satellite assemblies are a bummer... When discrete assembly files are
placed in the apk, they end up in `lib/{ABI}`, but that directory
mustn't have any subdirectories, so
`lib/{ABI}/{CULTURE}/assembly.dll.so` is out of question.  To work
around that, we need to mangle the satellite assembly name somehow and
that breaks image loader.  Will figure out how to fix this tomorrow.