[build] Simplify Visual Studio detection
CI recently started failing with:

    vswhere.exe result returned a directory that did not exist:
    System.IO.DirectoryNotFoundException: vswhere.exe result returned a directory that did not exist:
       at Xamarin.Android.Tools.VSWhere.MSBuildLocator.QueryLatest() in C:\a\_work\1\s\tools\vswhere\MSBuildLocator.cs:line 34

It is probably safe to assume that if vswhere can locate a Visual Studio
installation it will contain the core components that we depend on.  We
can simplify our vswhere invocation by always looking for and returning
the latest release or pre-release version of Visual Studio.