[Xamarin.Android.Build.Tasks] remove `@(_AndroidAssemblySkipCases)`
In an effort to remove old, unused code, we should be able to remove the
following related to the old Android.Support.* libraries:

* `Xamarin.Android.SkipCases.projitems` and the
  `@(_AndroidAssemblySkipCases)` item group.

This was a build performance optimization added in 02c07ed6 that would
*skip* a build step that normalized casing for `.dll` files with an
embedded `__AndroidLibraryProjects__.zip`. This optimization is not
needed at all for `.aar` files.

In .NET 9 projects, we don't create or embed
`__AndroidLibraryProjects__.zip`, so modern projects don't need this
optimization. Instead we bundle a `MyLibrary.aar` alongside
`MyLibrary.dll` inside of a `MyLibrary` NuGet package.

* Automatic trimming of the `Xamarin.Android.Support.*` assemblies

Modern packages mark themselves trimmable with `$(IsTrimmable)` or the
assembly-level attribute. Let's just remove the case for
`Android.Support` here.