[Xamarin.Android.Build.Tasks] remove `%(TrimMode)=link` metadata
To keep app sizes down in the .NET 6 timeframe, we special-cased
existing AndroidX and GPS packages so they were always trimmed.

Modern packages mark themselves trimmable with `$(IsTrimmable)` or the
assembly-level attribute.

For nearly 2 years, we've applied the appropriate trimming setting in
these packages:

* https://github.com/xamarin/AndroidX/pull/520
* https://github.com/xamarin/GooglePlayServicesComponents/pull/597

We should be able to remove this now in .NET 9.