Try to avoid `ObjectDisposedException` when writing output
It may happen that the contents stream in the
`WriteRequestContentToOutput` method is disposed while asynchronously
copying content to the connection output stream.  This will then cause
the `ObjectDisposedException` exception to be thrown when seeking back
to the beginning of the stream.

Try to avoid the problem by checking whether cancellation was requested
during the call to `CopyToAsync` and, if we get past that point, catch
the `ObjectDisposedException` to report it but not to rethrow it.