[Xamarin.Android.Build.Tasks] fix warning when missing `proguard-android.txt` (#8641)
We are seeing numerous test failures because of this warning.

    warning XA4304: ProGuard configuration file 'C:\Android\android-sdk\tools\proguard\proguard-android.txt' was not found.

This is because the `proguard-android.txt` file no longer ships with the
Android SDK. We should check this file exists before adding it to the
list of `@(_ProguardConfiguration)` files. This way if a user still does
have it, it will be included.