Attempt to fix assembly image load for typemap
This is for situations when, for any reason, Java is calling us before
any of the managed types involved in the mapping had the chance to be
loaded - that is, before their containing assembly has been loaded.

For most apps the overhead of this solution will be negligible (a single
`nullptr` check).