[WIP] This commit will break a LOT of tests
I need to change the assembly name mangling characters as, as it
appears, Android doesn't like `#` or `%` when deploying a Debug build
apk with `extractNativeLibs=true` in the manifest.  In such case, none
of the entries with `#` and `%` characters anywhere in the name (also
tested `!`) will be extracted to the filesystem.

TBC tomorrow, the great quest to find the Character That Works.