Soo.... Android, yes.
After a few months of names like `arc.so.bin` or `#Assembly.dll.so` or
`assemblies.arm64-v8a.blob.so` working just fine, Android 14 (on Pixel
6) stopped extracting entries with such names to the filesystem when the
`extractNativeLibs` flag is set to `true` in the manifest.  Not only
"special" characters (tested `#`, `!`, `~`, `@`, `%` and then gave up)
are no longer accepted - `lib/{ARCH]` entries with them are ignored and
not extracted to the file system, but also any entries without the `lib`
prefix are ignored (but ONLY on Release builds).

So, the new mangling style is:

  * `lib_` prefix for regular assemblies
  * `lib-culture-` prefix for satellite assemblies.

Also, all the entries we place in `lib` are now compressed, as per the
build mode.

Let's see which tests break now.