[Java.Interop] ignore remaining trimming warnings (#1190)
Fixes: https://github.com/xamarin/java.interop/issues/1157

The remaining trimmer warnings are around usage of System.Reflection
such as:

  * Creating an `Invoker` type from an original type via
    `Type.Assembly.GetType()`, within
    `JniRuntime.JniValueManager.GetInvokerType()`.

  * Create a `JniValueMarshaler` corresponding to
    `ParameterInfo.ParameterType` in
    `JniRuntime.JniMarshalMemberBuilder.GetParameterMarshaler()`.

  * Create a value corresponding to `ParameterInfo.ParameterType` in
    `ManagedPeer.GetValues()`.

  * Use `typeof(JavaObjectArray<>).MakeGenericType()` in
    `JniRuntime.JniTypeManager.GetTypes(JniTypeSignature)`.

  * Use `Type.MakeArrayType()` in
    `JniRuntime.JniTypeManager.GetTypes(JniTypeSignature)`.

The trimming warnings themselves all indicate that these types "might
be trimmed", in that nothing explicitly preserves them.  However, we
have a plethora of custom trimmer steps that work to preserve these
types involved in Java interop, such as:

  * https://github.com/xamarin/xamarin-android/blob/71b6fcc92f9e86246e0fe575bc1f44bbf73c3132/src/Microsoft.Android.Sdk.ILLink/PreserveApplications.cs
  * https://github.com/xamarin/xamarin-android/blob/71b6fcc92f9e86246e0fe575bc1f44bbf73c3132/src/Microsoft.Android.Sdk.ILLink/PreserveExportedTypes.cs
  * https://github.com/xamarin/xamarin-android/blob/71b6fcc92f9e86246e0fe575bc1f44bbf73c3132/src/Microsoft.Android.Sdk.ILLink/PreserveJavaExceptions.cs
  * https://github.com/xamarin/xamarin-android/blob/71b6fcc92f9e86246e0fe575bc1f44bbf73c3132/src/Microsoft.Android.Sdk.ILLink/PreserveJavaInterfaces.cs
  * https://github.com/xamarin/xamarin-android/blob/71b6fcc92f9e86246e0fe575bc1f44bbf73c3132/src/Microsoft.Android.Sdk.ILLink/PreserveRegistrations.cs

One day, in a perfect world, we could remove these trimmer steps and
completely rely on the trimmer's warnings & attributing mechanisms.
That day doesn't have to be today -- as our goal is begin surfacing
trimmer warnings to users in their own code and dependencies.

With these warnings ignored, we can fully enable analyzers with:

	<IsTrimmable>true</IsTrimmable>
	<EnableAotAnalyzer>true</EnableAotAnalyzer>

We can then remove:

	[assembly: AssemblyMetadata ("IsTrimmable", "True")]

As the MSBuild property does this for us, in addition to enabling
analyzers.

I don't think we should enable `$(IsAotCompatible)` quite yet, as
`Java.Interop.dll` likely *won't work* yet on NativeAOT, and this
places metadata that says an assembly is supported.  We should just
use the `$(EnableAotAnalyzer)` analyzers for now, so we don't
introduce new issues.